+++ requirements/frontend/src/components/friends/FriendsList.tsx
@@ -1,35 +1,10 @@
 import FriendRow from './FriendRow';
-import ConfirmRemoveModal from './modal/ConfirmRemoveModal';
-
 import '../../styles/friends_list.scss';
 import '../../styles/social_field.scss';
-import { FetchFriendsList } from '../../types/FetchFriendsList';
 import { useFriendList } from '../../Providers/FriendListProvider';
-import { useModal } from '../../Providers/ModalProvider';
 
 const FriendsList = () => {
-	const { setModalProps } = useModal();
 	const friendList = useFriendList();
-	const RemoveFriend = (friend: FetchFriendsList) => {
-		const userConfirm = async (confirmed: boolean) => {
-			if (confirmed) {
-				friendList.removeFriend(friend.id);
-			}
-			setModalProps(undefined);
-		};
-
-		setModalProps({
-			show: true,
-			content: (
-				<ConfirmRemoveModal
-					name={friend.displayName}
-					cb={userConfirm}
-				/>
-			),
-			width: '40%',
-			maxWidth: '500px',
-		});
-	};
 
 	if (friendList.friends.length === 0) {
 		return (
@@ -42,13 +17,7 @@ const FriendsList = () => {
 	return (
 		<ul className="friendslist-container">
 			{friendList.friends.map((friend) => {
-				return (
-					<FriendRow
-						key={friend.id}
-						onClick={() => RemoveFriend(friend)}
-						data={friend}
-					/>
-				);
+				return <FriendRow key={friend.id} data={friend} />;
 			})}
 		</ul>
 	);
